Basic Authentication Completed

+++ JournalApp/src/main/java/com/springbootProjects/JournalApp/Services/JournalEntityService/JournalEntityServices.java
@@ -4,7 +4,6 @@ import com.springbootProjects.JournalApp.Entity.JournalEntity.JournalEntity;
 import com.springbootProjects.JournalApp.Entity.UserEntity.UserEntity;
 import com.springbootProjects.JournalApp.Repository.JournalEntryRepository.JournalEntityRepository;
 import com.springbootProjects.JournalApp.Services.UserService.UserServices;
-import org.apache.catalina.User;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -33,7 +32,7 @@ public class JournalEntityServices
         UserEntity user= UserServices.findByName(userName); // here i am finding the user details as per the userName received
         JournalEntity saved=JournalEntityRepository.save(JournalEntity); // here i am saving the saved entries in variable
         user.getJournalEntries().add(saved); // means in UserEntity we have journalEntries there we are storing the data of journal entries
-        UserServices.addEntry(user); // we have updating the journalEntries in the UserEntity
+        UserServices.addUser(user); // we have updating the journalEntries in the UserEntity
     }
 
     // This is the overloaded method of  the addEntry where we are utilizing it for updating
@@ -52,13 +51,31 @@ public class JournalEntityServices
         return JournalEntityRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName)
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName)
     {
-        UserEntity user= UserServices.findByName(userName);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id)); // it matches the id from journalEntries i.e. from userEntity and id from JournalEntity
-        // if matches it removes the that id in UserEntity
-        // if we dont use this line the previous deleted record will be removed once we create the new entry, but maintain the consistency we are removing it manually
-        UserServices.addEntry(user);
-        JournalEntityRepository.deleteById(id);
+        boolean removed=false;
+        try
+        {
+            UserEntity user= UserServices.findByName(userName);
+             removed=user.getJournalEntries().removeIf(x-> x.getId().equals(id)); // it matches the id from journalEntries i.e. from userEntity and id from JournalEntity
+            // if matches it removes the that id in UserEntity
+            // if we dont use this line the previous deleted record will be removed once we create the new entry, but maintain the consistency we are removing it manually
+
+            if(removed)
+            {
+                UserServices.addUser(user);
+                JournalEntityRepository.deleteById(id);
+            }
+        }
+
+        catch(Exception e)
+        {
+            throw new RuntimeException("An error occurred while deleting the Journal Entry",e);
+        }
+
+        return removed;
+
+
     }
 }
